monto También era float, la puta madre.

+++ 2.C++/TPN8_Archivos/TPN8_EJ01.cpp
@@ -86,7 +86,7 @@ void CargarMontos(FILE *archivo,int &i) {
 }
 
 void ListarMontos(FILE *archivo) {
-    int monto;
+    float monto;
 
     rewind(archivo);
